travis: link against libssecomp 2.5.0 for tests
Travis now run tests with libssecomp 2.5.0 in the different
configurations (go 1.15.x, go 1.14.x, cgroup-systemd), except for the
fedora32 and centos7 Vagrant tests that still run with the old
libseccomp version.

This patch installs libssecomp 2.5.0 manually in the Travis host (for
the cgroup-systemd test) and in the Docker image (for the other tests).

Signed-off-by: Alban Crequy <alban@kinvolk.io>