Idmapped mounts: strengthen destination mount point creation
Take into consideration if the destination mount point is a file, folder and
if intermediate parts of the path have to be created too.

This commit also ignores the errors from mount_id_mapped as there are some
filesystems that don't support id mapped mounts yet and Kubernetes tries to
bind mount them (tmpfs in this case).